webapi/webapi-webspeech-w3c-tests/webspeech/blink/speech-synthesis-speak.html

<!--
Test convert from Blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/speechsynthesis/speech-synthesis-speak.html
-->

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Speech API Test: the basic mechanisms of speaking text work, including sending the job and receving the callback</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var t = async_test("Test that the basic mechanisms of speaking text work, including sending the job and receving the callback", { timeout: 5000});
t.step(function () {
  var u = new SpeechSynthesisUtterance(" ");

  u.onstart = function(event) {
    assert_greater_than(event.elapsedTime, 0);
    assert_true(speechSynthesis.speaking);
  }
  u.onend = function(event) {
    assert_greater_than(event.elapsedTime, 0);
    assert_false(speechSynthesis.speaking);
    t.done();
  }

  assert_false(speechSynthesis.speaking);
  speechSynthesis.speak(u);
});

</script>
